+++ stompbox/Stomps/Gain.h
@@ -11,17 +11,17 @@ enum
 class Gain : public StompBox
 {
 private:
-	double gain;
-	double minGain;
-	double maxGain;
-	double linearGain;
-	double currentLevel = 0;
+	float gain;
+	float minGain;
+	float maxGain;
+	float linearGain;
+	float currentLevel = 0;
 
 public:
-	Gain(double initialGain, double minGain, double maxGain);
+	Gain(float initialGain, float minGain, float maxGain);
 	~Gain() {}
 	virtual void init(int samplingFreq);
 	double GetLevel();
 
-	virtual void compute(int count, double* input, double* output);
+	virtual void compute(int count, float* input, float* output);
 };
